requirement.txt
streamlit
requests
beautifulsoup4
python-dotenv
sentence-transformers
faiss-cpu
langchain 
langchain-openai
langchain_huggingface
langchain-community
numpy
python-dotenv